Create path_utils submodule, search_jupyter_paths now returns generator
simplify imports
printing logic has been moved out of search_jupyter_paths and into the 
start of the apps
